+++ FE/json_facturas/DTE-01-0000MOO1-000000000000143.json
@@ -4,13 +4,13 @@
         "ambiente": "01",
         "tipoDte": "01",
         "numeroControl": "DTE-01-0000MOO1-000000000000143",
-        "codigoGeneracion": "25A5021B-646A-474C-A159-6CAD7E4B8EE4",
+        "codigoGeneracion": "7C43BF0B-F82C-43A0-9636-2224A85A6F12",
         "tipoModelo": 1,
         "tipoOperacion": 1,
         "tipoContingencia": null,
         "motivoContin": null,
         "fecEmi": "2025-03-07",
-        "horEmi": "15:59:14",
+        "horEmi": "20:58:26",
         "tipoMoneda": "USD"
     },
     "documentoRelacionado": null,
