+++ FE/json_facturas_firmadas/7c43bf0b-f82c-43a0-9636-2224a85a6f12.json
@@ -1,0 +1,4 @@
+{
+    "status": "OK",
+    "body": "eyJhbGciOiJSUzUxMiJ9.ew0KICAiaWRlbnRpZmljYWNpb24iIDogew0KICAgICJ2ZXJzaW9uIiA6IDEsDQogICAgImFtYmllbnRlIiA6ICIwMSIsDQogICAgInRpcG9EdGUiIDogIjAxIiwNCiAgICAibnVtZXJvQ29udHJvbCIgOiAiRFRFLTAxLTAwMDBNT08xLTAwMDAwMDAwMDAwMDE0MyIsDQogICAgImNvZGlnb0dlbmVyYWNpb24iIDogIjdDNDNCRjBCLUY4MkMtNDNBMC05NjM2LTIyMjRBODVBNkYxMiIsDQogICAgInRpcG9Nb2RlbG8iIDogMSwNCiAgICAidGlwb09wZXJhY2lvbiIgOiAxLA0KICAgICJ0aXBvQ29udGluZ2VuY2lhIiA6IG51bGwsDQogICAgIm1vdGl2b0NvbnRpbiIgOiBudWxsLA0KICAgICJmZWNFbWkiIDogIjIwMjUtMDMtMDciLA0KICAgICJob3JFbWkiIDogIjIwOjU4OjI2IiwNCiAgICAidGlwb01vbmVkYSIgOiAiVVNEIg0KICB9LA0KICAiZG9jdW1lbnRvUmVsYWNpb25hZG8iIDogbnVsbCwNCiAgImVtaXNvciIgOiB7DQogICAgIm5pdCIgOiAiMDYxNDI4MTEwMDEwNDAiLA0KICAgICJucmMiIDogIjEyNzg5MTMiLA0KICAgICJub21icmUiIDogIkludmVyc2lvbmVzIENvbWVyY2lhbGVzIEVzY29iYXIgU0EgZGUgQ1YiLA0KICAgICJjb2RBY3RpdmlkYWQiIDogIjQ2Mzc1IiwNCiAgICAiZGVzY0FjdGl2aWRhZCIgOiAiVmVudGEgYWwgcG9yIG1heW9yIGRlIHByb2R1Y3RvcyBsYWN0ZW9zIiwNCiAgICAibm9tYnJlQ29tZXJjaWFsIiA6ICJJbnZlcnNpb25lcyBDb21lcmNpYWxlcyBFc2NvYmFyIFNBIGRlIENWIiwNCiAgICAidGlwb0VzdGFibGVjaW1pZW50byIgOiAiMDIiLA0KICAgICJkaXJlY2Npb24iIDogew0KICAgICAgImRlcGFydGFtZW50byIgOiAiMDUiLA0KICAgICAgIm11bmljaXBpbyIgOiAiMTkiLA0KICAgICAgImNvbXBsZW1lbnRvIiA6ICJDYXJyZXRlcmEgYSBTb25zb25hdGUsIGttIDI3IMK9IFpvbmEgSW5kdXN0cmlhbCBMb3VyZGVzLCBDYWxsZSBBbnRpZ3VhIEhkYSwgTGEgTGliZXJ0YWQsIEVsIFNhbHZhZG9yIEMuQSINCiAgICB9LA0KICAgICJ0ZWxlZm9ubyIgOiAiNzc0NjIzMTAiLA0KICAgICJjb3JyZW8iIDogImluZm9ybWF0aWNhQGdydXBvaW5jb2UuY29tIiwNCiAgICAiY29kRXN0YWJsZU1IIiA6ICJNMDAxIiwNCiAgICAiY29kRXN0YWJsZSIgOiAiMDAwMSIsDQogICAgImNvZFB1bnRvVmVudGFNSCIgOiAiUDAwMSIsDQogICAgImNvZFB1bnRvVmVudGEiIDogIjAwMDEiDQogIH0sDQogICJyZWNlcHRvciIgOiB7DQogICAgIm5vbWJyZSIgOiAiRnJhbmNpc2NvIEFudG9uaW8gRmxvcmVzIiwNCiAgICAiY29kQWN0aXZpZGFkIiA6ICIyNDMxMCIsDQogICAgImRlc2NBY3RpdmlkYWQiIDogInVuZGljacOzbiBkZSBoaWVycm8geSBhY2VybyIsDQogICAgImRpcmVjY2lvbiIgOiB7DQogICAgICAiZGVwYXJ0YW1lbnRvIiA6ICIwNSIsDQogICAgICAibXVuaWNpcGlvIiA6ICIxOSIsDQogICAgICAiY29tcGxlbWVudG8iIDogIkNhcnJldGVyYSBhIFNvbnNvbmF0ZSwga20gMjcgwr0gWm9uYSBJbmR1c3RyaWFsIExvdXJkZXMsIENhbGxlIEFudGlndWEgSGRhLCBMYSBMaWJlcnRhZCwgRWwgU2FsdmFkb3IgQy5BIg0KICAgIH0sDQogICAgInRlbGVmb25vIiA6ICI2MDQ0Nzc0MzUiLA0KICAgICJjb3JyZW8iIDogImp1bmlvcmZyYW5AaG90bWFpbC5lcyIsDQogICAgIm5yYyIgOiBudWxsLA0KICAgICJ0aXBvRG9jdW1lbnRvIiA6ICIzNiIsDQogICAgIm51bURvY3VtZW50byIgOiAiMDYxNDI4MDI5NjEwNjYiDQogIH0sDQogICJvdHJvc0RvY3VtZW50b3MiIDogbnVsbCwNCiAgInZlbnRhVGVyY2VybyIgOiBudWxsLA0KICAiY3VlcnBvRG9jdW1lbnRvIiA6IFsgew0KICAgICJudW1JdGVtIiA6IDEsDQogICAgInRpcG9JdGVtIiA6IDEsDQogICAgIm51bWVyb0RvY3VtZW50byIgOiBudWxsLA0KICAgICJjb2RpZ28iIDogIlBDSDQwNSIsDQogICAgImNvZFRyaWJ1dG8iIDogbnVsbCwNCiAgICAiZGVzY3JpcGNpb24iIDogImxlY2hlIGVudGVyYSBpcmEgMjYgMjUwZ2IiLA0KICAgICJjYW50aWRhZCIgOiAxLjAsDQogICAgInVuaU1lZGlkYSIgOiA1OSwNCiAgICAicHJlY2lvVW5pIiA6IDguNSwNCiAgICAibW9udG9EZXNjdSIgOiAwLjAsDQogICAgInZlbnRhTm9TdWoiIDogMC4wLA0KICAgICJ2ZW50YUV4ZW50YSIgOiAwLjAsDQogICAgInZlbnRhR3JhdmFkYSIgOiA4LjUsDQogICAgInRyaWJ1dG9zIiA6IG51bGwsDQogICAgInBzdiIgOiAwLjAsDQogICAgIm5vR3JhdmFkbyIgOiAwLjAsDQogICAgIml2YUl0ZW0iIDogMC45OA0KICB9IF0sDQogICJyZXN1bWVuIiA6IHsNCiAgICAidG90YWxOb1N1aiIgOiAwLjAsDQogICAgInRvdGFsRXhlbnRhIiA6IDAuMCwNCiAgICAidG90YWxHcmF2YWRhIiA6IDguNSwNCiAgICAic3ViVG90YWxWZW50YXMiIDogOC41LA0KICAgICJkZXNjdU5vU3VqIiA6IDAuMCwNCiAgICAiZGVzY3VFeGVudGEiIDogMC4wLA0KICAgICJkZXNjdUdyYXZhZGEiIDogMC4wLA0KICAgICJwb3JjZW50YWplRGVzY3VlbnRvIiA6IDAuMCwNCiAgICAidG90YWxEZXNjdSIgOiAwLjAsDQogICAgInN1YlRvdGFsIiA6IDguNSwNCiAgICAiaXZhUmV0ZTEiIDogMC4wLA0KICAgICJyZXRlUmVudGEiIDogMC4wLA0KICAgICJ0b3RhbE5vR3JhdmFkbyIgOiAwLjAsDQogICAgInRvdGFsTGV0cmFzIiA6ICJPY2hvIGNvbiA1MC8xMDAgVVNEIiwNCiAgICAic2FsZG9GYXZvciIgOiAwLjAsDQogICAgImNvbmRpY2lvbk9wZXJhY2lvbiIgOiAxLA0KICAgICJwYWdvcyIgOiBudWxsLA0KICAgICJudW1QYWdvRWxlY3Ryb25pY28iIDogbnVsbCwNCiAgICAidHJpYnV0b3MiIDogbnVsbCwNCiAgICAidG90YWxJdmEiIDogMC45OCwNCiAgICAibW9udG9Ub3RhbE9wZXJhY2lvbiIgOiA4LjUsDQogICAgInRvdGFsUGFnYXIiIDogOC41DQogIH0sDQogICJleHRlbnNpb24iIDogew0KICAgICJub21iRW50cmVnYSIgOiBudWxsLA0KICAgICJkb2N1RW50cmVnYSIgOiBudWxsLA0KICAgICJub21iUmVjaWJlIiA6IG51bGwsDQogICAgImRvY3VSZWNpYmUiIDogbnVsbCwNCiAgICAib2JzZXJ2YWNpb25lcyIgOiAiIiwNCiAgICAicGxhY2FWZWhpY3VsbyIgOiBudWxsDQogIH0sDQogICJhcGVuZGljZSIgOiBudWxsDQp9.a2T_iIlqQTnKaxuKX9SYxUSOr9XF-WwXaONcTaoWDZ2HoF_uK_GDwq0Tt6JeWEQFWKdjTuXSFQHQuHaLzKWhkIboWFDNM-ygqOvyCGMy08yLJRokBV3H5HA0WyDRRo5NAHhbd-0Rd6D7_jAQq-esdTNZj7lfZ37vEyHS2-d8sRQKnv9QIunEwaoP_aUpNaU7k4Eq3TQ1Mb5E6bU7uv_gQyMsQQazDmkCCUTTVd8p0gLOP7nyKt8yDSnjQ98f6F1OHJIrCcWdqRK54sLVQBeYMK4D_I_HnCnmE3f8ZnVdggwa1BtKRMyf95qkU_fPTqWP5NSD9NSDMc9sCguBWsZN5w"
+}
