+++ intracoe/urls.py
@@ -11,7 +11,7 @@ urlpatterns = [
     path('fe/', include('FE.urls')),
     path('contabilidad/', include('CONTABILIDAD.urls')),
     path('informatica/', include('INFORMATICA.urls')),
-    path('dte/', include('FE.dte.urls')),
+    #path('dte/', include('FE.dte.urls')),
 
     path('login/', auth_views.LoginView.as_view(), name='login'),
     path('logout/', auth_views.LogoutView.as_view(), name='logout'),
