--- json_facturas/DTE-01-0000M001-000000000000012.json
@@ -1,108 +1,0 @@
-{
-    "identificacion": {
-        "version": 1,
-        "ambiente": "01",
-        "tipoDte": "01",
-        "numeroControl": "DTE-01-0000M001-000000000000012",
-        "codigoGeneracion": "99BA6C02-44F8-4201-9BD1-0B8B17A715E5",
-        "tipoModelo": 1,
-        "tipoOperacion": 1,
-        "tipoContingencia": null,
-        "motivoContin": null,
-        "fecEmi": "2025-02-15",
-        "horEmi": "07:48:03",
-        "tipoMoneda": "USD"
-    },
-    "documentoRelacionado": null,
-    "emisor": {
-        "nit": "06142811001040",
-        "nrc": "05182",
-        "nombre": "Inversiones Comerciales Escobar SA de CV",
-        "codActividad": "10501",
-        "descActividad": "Fabricación de productos lácteos excepto sorbetes, quesos y sucedáneos no lácteos o sustitutos",
-        "nombreComercial": "Inversiones Comerciales Escobar SA de CV",
-        "tipoEstablecimiento": "02",
-        "direccion": {
-            "departamento": "05",
-            "municipio": "19",
-            "complemento": "Carretera a Sonsonate, km 27 ½ Zona Industrial Lourdes, Calle Antigua Hda, La Libertad, El Salvador C.A"
-        },
-        "telefono": "77462310",
-        "correo": "informatica@grupoincoe.com",
-        "codEstableMH": "M001",
-        "codEstable": "0001",
-        "codPuntoVentaMH": "P001",
-        "codPuntoVenta": "0001"
-    },
-    "receptor": {
-        "tipoDocumento": "36",
-        "numDocumento": "06142802961066",
-        "nrc": null,
-        "nombre": "Francisco Antonio Flores",
-        "codActividad": "24310",
-        "descActividad": "undición de hierro y acero",
-        "direccion": {
-            "departamento": "05",
-            "municipio": "19",
-            "complemento": "Carretera a Sonsonate, km 27 ½ Zona Industrial Lourdes, Calle Antigua Hda, La Libertad, El Salvador C.A"
-        },
-        "telefono": "77462310",
-        "correo": "juniorfran@hotmail.es"
-    },
-    "otrosDocumentos": null,
-    "ventaTercero": null,
-    "cuerpoDocumento": [
-        {
-            "numItem": 1,
-            "tipoItem": 1,
-            "numeroDocumento": null,
-            "codigo": "PCH405",
-            "codTributo": null,
-            "descripcion": "leche entera ira 26 250gb",
-            "cantidad": 1.0,
-            "uniMedida": 59,
-            "precioUni": 10.0,
-            "montoDescu": 0.0,
-            "ventaNoSuj": 0.0,
-            "ventaExenta": 0.0,
-            "ventaGravada": 10.0,
-            "tributos": null,
-            "psv": 0.0,
-            "noGravado": 0.0,
-            "ivaItem": 1.3
-        }
-    ],
-    "resumen": {
-        "totalNoSuj": 0.0,
-        "totalExenta": 0.0,
-        "totalGravada": 10.0,
-        "subTotalVentas": 10.0,
-        "descuNoSuj": 0.0,
-        "descuExenta": 0.0,
-        "descuGravada": 0.0,
-        "porcentajeDescuento": 0.0,
-        "totalDescu": 0.0,
-        "tributos": null,
-        "subTotal": 10.0,
-        "ivaRete1": 0.0,
-        "reteRenta": 0.0,
-        "montoTotalOperacion": 11.3,
-        "totalNoGravado": 0.0,
-        "totalPagar": 11.3,
-        "totalLetras": "Once con 30/100 USD",
-        "totalIva": 1.3,
-        "saldoFavor": 0.0,
-        "condicionOperacion": 1,
-        "pagos": null,
-        "numPagoElectronico": null
-    },
-    "extension": {
-        "nombEntrega": null,
-        "docuEntrega": null,
-        "nombRecibe": null,
-        "docuRecibe": null,
-        "observaciones": null,
-        "placaVehiculo": null
-    },
-    "apendice": null
-}
