--- json_facturas/DTE-01-M001-000000000000004.json
@@ -1,120 +1,0 @@
-{
-    "identificacion": {
-        "version": 2,
-        "ambiente": "01",
-        "versionApp": 2,
-        "tipoDte": "01",
-        "numeroControl": "DTE-01-M001-000000000000004",
-        "codigoGeneracion": "8337E27C-74CB-447D-B696-605A07FCF40B",
-        "tipoModelo": 1,
-        "tipoOperacion": "1",
-        "documentoRelacionado": "",
-        "otrosDocumentos": [],
-        "ventaTercero": "",
-        "apendice": "",
-        "tipoContingencia": "00",
-        "motivoContin": "NA",
-        "fecEmi": "2025-02-14",
-        "horEmi": "09:00:38",
-        "tipoMoneda": "USD"
-    },
-    "emisor": {
-        "nit": "06142811001040",
-        "nrc": "05182",
-        "nombre": "Inversiones Comerciales Escobar SA de CV",
-        "codActividad": "10501",
-        "descActividad": "Fabricación de productos lácteos excepto sorbetes, quesos y sucedáneos no lácteos o sustitutos",
-        "nombreComercial": "Inversiones Comerciales Escobar SA de CV",
-        "tipoEstablecimiento": "02",
-        "codEstable": "M001",
-        "codPuntoVenta": "P001",
-        "direccion": {
-            "departamento": "Otro Pais",
-            "municipio": "Otro Pais",
-            "complemento": "Carretera a Sonsonate, km 27 ½ Zona Industrial Lourdes, Calle Antigua Hda, La Libertad, El Salvador C.A"
-        },
-        "telefono": "77462310",
-        "correo": "informatica@grupoincoe.com",
-        "codEstableMH": "M001",
-        "codPuntoVentaMH": "P001"
-    },
-    "receptor": {
-        "tipoDocumento": "13",
-        "numDocumento": "053208681",
-        "nombre": "Francisco Antonio Flores",
-        "nrc": "05182",
-        "direccion": {
-            "departamento": "Otro Pais",
-            "municipio": "Otro Pais",
-            "complemento": "Carretera a Sonsonate, km 27 ½ Zona Industrial Lourdes, Calle Antigua Hda, La Libertad, El Salvador C.A"
-        },
-        "telefono": "77462310",
-        "correo": "juniorfran@hotmail.es"
-    },
-    "cuerpoDocumento": [
-        {
-            "numeroDocumento": "1",
-            "tipoItem": 1,
-            "codigo": "PCH405",
-            "descripcion": "leche entera ira 26 250gb",
-            "cantidad": 1,
-            "uniMedida": "UND",
-            "precioUni": 10.0,
-            "montoDescu": 0.0,
-            "ventaExenta": 0.0,
-            "ventaNoSuj": 0.0,
-            "ventaGravada": 10.0,
-            "ivaItem": 1.3,
-            "codTributo": "01",
-            "psv": "0",
-            "noGravado": "0",
-            "tributos": [
-                {
-                    "codTributo": "01",
-                    "montoTributo": 0.0,
-                    "porcentajeTributo": 0.0,
-                    "tipoTributo": "IVA"
-                }
-            ]
-        }
-    ],
-    "resumen": {
-        "totalGravada": 10.0,
-        "totalExenta": 0.0,
-        "totalNoSuj": 0.0,
-        "subTotalVentas": 10.0,
-        "descuNoSuj": 0.0,
-        "descuExenta": 0.0,
-        "descuGravada": 0.0,
-        "porcentajeDescuento": 0.0,
-        "totalDescu": 0.0,
-        "tributos": [
-            {
-                "codigo": "01",
-                "descripcion": "IVA",
-                "valor": 0.0
-            }
-        ],
-        "subTotal": 10.0,
-        "totalPagar": 11.3,
-        "totalLetras": "Once con 30/100 USD",
-        "condicionOperacion": "01",
-        "pagos": [
-            {
-                "codigo": "001",
-                "referencia": "REF001",
-                "periodo": "2025-02",
-                "plazo": 30
-            }
-        ],
-        "numPagoElectronico": "0000"
-    },
-    "extension": {
-        "nombEntrega": "NO APLICA",
-        "docuEntrega": "0000000000",
-        "nombRecibe": "NO APLICA",
-        "docuRecibe": "0000000000",
-        "placaVehiculo": "XXX000",
-        "observaciones": ""
-    }
-}
